requirements.txt
taichi
numpy
PyOpenGL
glfw
trimesh
scikit-image

